+++ Changelog.md
@@ -7,6 +7,7 @@ v2.11.0 (not published yet)
 - Add new macros "random:digits", "current_date", "current_time", "current_time24" ([#802](/../../issues/802))
 - Fix UI issues regarding mkvextract
 - Fix remove confirmation on Lists showing first item name only even if multiple items are selected
+- Add NVEnc parameters --lut3d, --lut3d_interp, --dolby-vision-rpu, --dolby-vision-profile
 - Update tools
   - MKVToolNix v66.0
   - NVEnc v5.46